+++ CI/SCRIPTS/script_utils.py
@@ -22,7 +22,7 @@ from functools import wraps
 
 from sertit import AnyPath, s3, unistra
 from sertit.misc import ListEnum
-from sertit.unistra import UNISTRA_S3_ENPOINT
+from sertit.unistra import UNISTRA_S3_ENDPOINT
 
 CI_SERTIT_S3 = "CI_SERTIT_USE_S3"
 
@@ -113,6 +113,6 @@ def xml_path():
 
 
 def s3_env(function):
-    return s3.s3_env(default_endpoint=UNISTRA_S3_ENPOINT, use_s3_env_var=CI_SERTIT_S3)(
+    return s3.s3_env(default_endpoint=UNISTRA_S3_ENDPOINT, use_s3_env_var=CI_SERTIT_S3)(
         function
     )