+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.42.1 (2024-09-dd)
+## 1.42.1 (2024-09-25)
 
 - FIX: Don't use isinstance and subtypes together ([#22](https://github.com/sertit/sertit-utils/issues/22))
 - FIX: Allow `use_s3_env_var` in `unistra.s3_env`